% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afFILLiator.R
\name{afFILLiator}
\alias{afFILLiator}
\title{Get authors name, affiliations and PMID from where affiliation was gathered}
\usage{
afFILLiator(fore_name = "Benjamin", last_name = "Valderrama")
}
\arguments{
\item{fore_name}{A character vector of one element with the name(s) of the author.}

\item{last_name}{A character vector of one element with the lastname of the author.}
}
\value{
A dataframe with author name, lastname, affiliations and the PMID of the publication where the information was gather from.
}
\description{
Get authors name, affiliations and PMID from where affiliation was gathered
}
\examples{
afFILLiator(fore_name = "John F", last_name = "Cryan")
}
