% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afFILLiator.R
\name{fetch_record}
\alias{fetch_record}
\title{Get a XML object with the article information from a PMID}
\usage{
fetch_record(pmid)
}
\arguments{
\item{pmid}{a numeric vactor of one element with the PMID.}
}
\value{
an XML document with the article information associated to that PMID.
}
\description{
Get a XML object with the article information from a PMID
}
\examples{
pmid <- 38852762
fetch_record(pmid)
}
